poller: Add (and export to other C code) two protected funcs
While playing with HTTP code, I noticed that libevent HTTP code
needs to access the event_base and the evdns_base structures.

So, here's a commit that saves these two structures into the poller
and that adds "protected" functions to access such structures.

The two functions that I added are protected in the sense that one
can access them from the C code of libneubot, but they are not
accessible from Python and Java, because they are not mentioned
in the libneubot.json API interface.