poller: return status_t in defer_read and defer_write
This commit changes defer_read and defer_write in poller.c to return
a status_t rather than a NeubotEvent structure.

As a result, the NeubotEvent_cancel method is no longer needed.

I am making this change, because I don't need the cancel functionality
and also because I need a simpler API to implement the automatic
wrapping of the C classes more easily.