libneubot.json: add hooks, slots and __destructor__
1) Add to the type model two new types: the hook and the slot.

The hook represents a state-less slow operation that invokes
a callback of an object when the event occurs.

For example, defer_read receives ["hook_vo, ["handle_ok",
"handle_timeout"]], which in C is expanded as follows:

    ... neubot_hook_vo handle_ok, neubot_hook_vo handle_timeout,
        void *opaque

As you see, when the hook is expanded in C, we add the opaque
object to which the callbacks make reference.

The slot is an hook that is the argument of a constructor of
an object, which allows you to extend the basic C object by
telling it what to do when some event occur.

The difference between a slot and a hook is in the implementation:

The slot, in fact, saves the context in the constructor of the
object, therefore the saved context and the object have the same
life cycle.

Conversely, the state of the hook is a closure (which includes
the functions(s) to be called and the opaque pointer) that is
destroyed after the hook is called.

By the way, hook_vo means "void hook that receives an opaque
object", and hook_vos indicates that the hook also receives
a second argument that is a C string.

I added the hooks and the slots, because they helped me to
implement more easy to use wrappers for Python, which I will
commit in one or two commits.

2) Add the __destructor__ special function, which indicates
that a function is a destructor.  This feature, again, is
needed to implement better Python wrappers.