Makefile.in: install also log.h
No need to crank major, because we already cranked it due to
the changes in the Pollable constructor and attach().

This diff was originally meant to go in the next release (and is
preparatory for installing more headers), however, it was anticipated
to take advantage of the cranking of the major number.

While there, reorganize the way in which the sources and headers
are listed in the Makefile.in file, to reduce the likelyhood of
generating a merge conflict.