Rewrite the glue between C++ and C code
I was not sure whether the static_cast<ChildClass *>(parent)
strategy was OK memory-wise.

Well, it turns out it was not OK. Sure, the new pointer was for
an instance of ChildClass, but the program was reading from space
that was not allocated (according to Valgrind).

Therefore, change strategy: use the constructors as they are
typically used in C++, live with exceptions (in any case,
Alessandro confirmed that we compile libneubot with exceptions
under Android), use try...catch as appropriate.