gen_python: simplify hook and slots, remove unneeded glue
This basically removes the wrapper function calls that used
to raise an exception on error.

Because we use a C/C++ event loop, it is not pratical to
raise exceptions when we use libneubot, because in the
worst case they are catched by ctypes and not propagated
eventually to unwind the whole stack.

Update the regression tests accordingly.