Implement __gettable__ / __object_getter__
This new couple of markers for libneubot.json allows us to define
objects (passed to the constructor) that one can get.

This way, when we are in Python and we receive a `poller` Python
object, we can save such object and return it in the proper getter
method rather than returning the low level object (i.e., a pointer)
that the `poller` Python object wraps.

To do so, we must mark with `__gettable__` the `poller` passed to
the constructor, and we must mark with `__object_getter__` the getter
function that retrieves the poller.