Remove wrappers from C++ to C and Python
The wrappers from C++ to C and Python were added to LibNeubot following
this rationale:

1) LibNeubot was to be written in C++ for SWIG to generate more
   easily usable NDK/JNI code;

2) The C interface was helpful to cherry-pick functions from Python
   using ctypes and to create a nice Python interface around the rought
   edges of the ctypes interface. In turn, this was because LibNeubot
   was to be used both for mobile and Neubot for PC.

After the discussion with @hellais that led to Libight, the scope was
narrowed to include only 1), because Libight is meant to support Neubot
and OONI on mobile platforms only.

Therefore (and, also, because the wrappers are on my way now that I am
working on a fully C++ poller that uses std::function and other C++11
goodies), I suggest that we remove the `optional complexity` represented
by 2) to develop more quickly.

In fact, if 2) is not a Libight objective, my time is better spent working
on the C++ core rather than working to keep 2) up to date.

I tested this change by rebuilding everything as in `make distclean;
autoreconf -i; ./configure; make V=0 check', without breakages or new
warnings.

I am aware that this diff removes some Python based tests, but C++ tests
are to be written in any case (I'd rather do testing of the C++ core than
do testing of the Python wrapper of the C wrapper of the core itself).