Make the LICENSE more community friendly
The formulation of the license with 'each severally entitled'
works best if only Arturo and I are authors, and we plan to
receive no contributions from other people.

However, IMHO such sentence -- which implements disjoint
ownership -- also discourage contributions.

On the one hand, it is not common to see such sentence into the
license of a software. If it is strange, it raises eyebrows.
If it is not smooth, it is do-something-else.

On the other hand, the community should based on a pact by which
contributors can be sure that they contributed to a commons put
under the BSD license. To implement this, shared property (rather
than disjoint property) is needed.

So, I removed the disjoint property sentence and I extended
the copyright to include all the contributors.