configure.ac: check whether C++ lib supports C++11
We've started to use many C++11 features with the new Poller, e.g.,
std::function and std::move, so better be sure that we are on a system
with a C++11 compliant standard library.

While there, improve the error case of the check whether the
C++ compiler knows the C++11.