poller: don't always register a SIGINT handler
A library should provide mechanism, not implement policy. To break
on SIGINT is policy. It would be mechanism to allow the library user
to register a callback to handle SIGINT.

Since we don't plan to handle SIGINT on Android, this functionality
seems to be only needed for implementing test programs.

Therefore, do not bother with writing the policy and only provide
the minimum functionality needed for a test program.

In particular, this diff moves the code to register and unregister
the SIGINT handler into the `break_loop_on_sigint_()` method.

I've given this method a name terminated by underscore to stress
that it's not meant to be used outside of testing programs.

For the same reason, I've not added a `ight_break_loop_on_sigint()`
top level function to libight.