poller: fail, rather than abort, DNS requests on exit
Maybe this is just paranoia, but I'd rather fail than abort
the pending DNS requests.

Let's assume that there is a class that saves a pointer to
a pending DNS request, or NULL if the request completed.

In this scenario, suppose that the poller is destroyed and
then the destructor of such class is run.

If the DNS request was aborted in the poller, the class has
no way to know that its pointer is no longer valid.

Conversely, if the request is failed in the poller and the
class is correctly written, the DNS callback is invoked and
the class has a chance to NULL its pointer.