Do not rely on global objects because they don't work in static libs
Basically, it is bad practice to rely on global objects, because their
constructor is only called when they are embedded into a dyamic lib.

I discovered this problem when I compiled libight in static mode to
make sure that there were no leaks in the tests using valgrind.

When libight is compiled in static mode, in particular, all the functions
of IghtLibevent are not properly initialized, leading to std::terminate.

To fix this problem, it suffices to use the singleton pattern [1] to
allocate a global-like IghtPoller and IghtLibevent.

[1] http://stackoverflow.com/questions/1008019/c-singleton-design-pattern