libevent.h: add IghtEvbuffer wrapper for evbuffer
The IghtEvbuffer class wraps for libevent's evbuffer. It provides move
semantic. It lazy-allocates an underlying evbuffer only when needed.

I plan to use this class (or a more complex class based on this class) in
connection.{cpp,h} to implement RAII and, thus, to simplify the constructor
and destructor of the IghtConnection class.