connection: remove the buffered read/write code
In libneubot the connection needed to implement buffering both for
reading and for writing. In fact, libneubot idea was to implement
buffering efficiently in C/C++ and to implement higher level stuff
either in Python or in Java.

This is no longer the case. Libight is fully written in C++. So
we can move the buffering outside IghtConnection.

Specifically, this commit removes the buffering. In another pull
request, I will take the code removed here and use it to implement
a buffer class as a wrapper of `evbuffer`.