Remove connection's factory methods
As discussed with @hellais, the factory methods need to go.

We discussed the possiblity of having a do-nothing constructor,
coupled with attach() and connect().

Since there is the possibility to provide a default value for
variables, however, I've decided to have two constructors, one
for attaching to a socket, the other for connecting.

I prefer this solution because it leads to a simpler API
and, therefore, to less lines of code.

Compare:

    auto conn1 = new IghtConnection(so);
    auto conn2 = new IghtConnection("PF_INET", "www.google.com", "80");

with:

    auto conn1 = new IghtConnection();
    auto conn2 = new IghtConnection();

    conn1->attach(so);
    conn2->connect("PF_INET", "www.google.com", "80");