.gitignore: revert changes to make smooth merge possible
The changes that this branch applies to .gitignore prevent a smooth merge
of this branch with `add-libevent-abstraction-layer-step1`.

Since the change is small, and since solving conflicts manually is a
bit annoying, revert the change done by this branch to .gitignore.

I will reapply the changes once this branch is merged.