Revert more changes to make concurrent work easier
There is a lesson in here: if I make too many changes within a branch, I
cannot work concurrently on other stuff.

In this case, in particular, I cannot work to add the new buffer code and
the regression tests, because this conflicts with me moving the test program
for the connection inside the `test/net` folder.