net: stop using `long long` to represent sockets
It was a libneubot-ism to use `long long` to represent sockets,
since it was the most handy way to exchange sockets between Java
and native code (i.e., using jlong).

In libight, we can safely use evutil_socket_t.

I'm not going to delete the code in net/ll2sock.h, however. I
think it's useful to have macros and inline functions to ensure
that the sockets we are moving around are in range.