Add amalgamated JsonCpp sources
Also add the very simple script used to generate the
amalgamated sources, amalgamate.sh.

I opted for adding the amalgamated source file and
header in the repository, as opposed to automatically
generating it at compile time, for now.

If we see that autogenerating the file is better, we
can switch to that model later.