dns: tests: replace throw runtime_error with safer idiom
The idiom by which we throw a runtime_error when a test fails is
broken because libevent's loop does not know about exceptions.

So, when an exception is thrown, we exit from the loop without the
loop knowing that we exited.

Therefore, next time we invoke the loop libevent complains since
it thinks that we are nesting loops.

As a consequence, all subsequent test fails because we cannot
re-enter into libevent's loop.

To fix this, use a safer idiom: when a callback that was not expected
to be called is called, set a boolean to true, break the loop, and
then check after the loop whether the boolean is true or not.

While there, also avoid throwing exception in the cancel-is-idempotent
tests, where we don't need to signal any error in the delayed call.