dns: tidy up DNSResolver() invocation using DNSSettings()
Rather than having too many params in DNSResolver() constructor, let's
roll out a class for holding the params.

Has the advantage that you only need to set the values you need.

It would have been much better if C++ supported designated initializers
such that the following would have been acceptable:

    auto r = ight::DNSResolver({
        .timeout = 0.5,
        .nameserver = "8.8.8.8"
    });

However, designated initializers are a C99 only feature, so I'm not
so comfortable with using them in libight (yes, clang and gcc accept
this extension, but maybe in the future we'll need another compiler
that runs on some strange system).

So, what we have instead is the following:

    auto r = ight::DNSResolver(ight::DNSSettings()
        .set_timeout(0.5).set_nameserver("8.8.8.8"));

More boring, but for now it would do.