Make DNSResolver() lazy
Solves the design issue by which DNSResolver was an alias for the default
evdns resolver when no settings were specified, a different beast otherwise.

This will be needed by the next patch that will disable by default the 0x20
hack for the DNSResolver, under the motivation that in OONI we want more
control and we want to enable 0x20 explicitly.

Without the change of this patch, DNSResolver() would have the 0x20 disabled
when constructed with parameters and enabled otherwise.

This is clearly stupid and annoying.

Especially because the reason for having the without-parameters constructor
doing a different thing was to avoid allocation during move semantic.

So, let's make DNSResolver lazy and let's avoid future surprises.