DNSResolver, DNSSettings: use negative values to say "don't"
One may want to specify zero as a parameter, even if this is a stupid
thing to do, and even if I don't know if libevent supports that.

Just allow that in our interface.

Of course this requires to switch from unsigned to signed (yes, this
second part was fixed-up into the original patch ;-).