Bind a resolver with its requests
Strengthen the ties between a resolver and the requests
generated by such resolver.

Now, requests have the same lifecycle as the resolver, and
one does not need to keep track of them.

Actually, one does not even see them, as they are managed
by the resolver under the hood.

Reduces the amount of pain generated by our DNS API.