Fix use-after-free in protocols/dns.cpp
The previous implementation caused an use-after-free when the
callback of a request replaced the Request object which caused
RequestImpl to be destroyed (see d931724c0ab...).

Unfortunately, after the callback returned, there was still
code using the just-deleted RequestImpl object.

More generally, the problem was caused by the programming pattern
used, which deleted RequestImpl from different code paths.

So, I have chosen to fix this bug changing the code pattern
used, to simplify future changes to the code.

Now, RequestImpl always deletes itself. To know when its controlling
Request object is dead, Request and RequestImpl communicate through
a shared pointer to a bool object.

Such shared boolean is checked by RequestImpl to decide whether
to call the callback, or whether this shall not be done.

Request sets the shared boolean to false when its destructor is
called, thereby telling RequestImpl to ignore the callback.

The shared boolean controlled through a shared pointer, so it is
deleted only when both Request and RequestImpl die.

After this change, valgrind does not show any memory error
when running the DNS regress test.