Make IghtDelayedCall Non{Copy,Mov}able
Since IghtDelayedCall contains lambdas bound to its `this` pointer,
it should not be moved or copied. Use SharedPointer<> instead.

Note that the implementation may be simplified to avoid allocating
a pointer to a std::function; not doing this here, because it would
complicate the patch and make it less obvious.

This diff also propagates changes so that the tree compiles again,
basically using SharedPointer<> in many cases and avoiding move semantic
in other cases.