Make IghtStringVector non {copy,mov}able
It contains pointers to objects allocated using malloc, so it
should not be copied and/or moved using the default implementation
of copy and move. Rather make it non copyable and non movable so
one can copy or move it through a SharedPtr<>.

By the way, this object is only used by connection and is a
legacy of when libneubot was to be implemented using a minimal
C++ subset, therefore I guess I should switch connection to
use a std::vector<> and remove this IghtStringVector.