dns::Response can safely be copied
In fact, all the fields are copyable and movable, and there is no harm
in copying and/or in moving them.

Update the regress test. The default C++ behavior is to copy integers
and other simple types, rather than move.

The fact that the integer members of a response are copied rather than
moved when move semantic is used should not be a problem.

Also because move semantic is typically used when one is not interested
into the fields of the moved object after the move itself.