Incomplete but explanatory bugfix
Basically, if Stream is created on the stack and then moved, `this`'s
value changes every time we move the object.

Unfortunately, lambdas (such as connection.on_error() set in the
constructor) are bound to the `this` used when they were instantiated.

Since Request moves stream once and since on_error() is set in the
constructor, the `this` to which on_error() argument is bound is gone
when we try to access it, hence the crash.

The committed fix is only enough to pass the test and to provide a proof
and an explanation.

The right thing to do, instead, is to always allocate on the heap and
manage through smart pointers object bound to lambdas.