Proper bugfix: allocate on heap, use smart pointer
This is the proper bugfix for the http::Stream memory corruption
bug that was exposed a few commits ago.

Allocate http::Stream always on the heap. Make it non movable (was
only non copyable) because its `this` is bound to lambdas.

Manage its lifecycle through smart pointers.

A little more syntactic sugar will come (not in this branch
that is about fixing stuff).

I've yet to check whether the same bug is present in other classes
of the tree.

After this commit:

==30795== Memcheck, a memory error detector
==30795== Copyright (C) 2002-2013, and GNU GPL'd, by Julian Seward et al.
==30795== Using Valgrind-3.10.0 and LibVEX; rerun with -h for copyright info
==30795== Command: ./test/protocols/http
==30795==

[snip]

==30795== HEAP SUMMARY:
==30795==     in use at exit: 0 bytes in 0 blocks
==30795==   total heap usage: 2,521 allocs, 2,521 frees, 489,485 bytes allocated
==30795==
==30795== All heap blocks were freed -- no leaks are possible
==30795==
==30795== For counts of detected and suppressed errors, rerun with: -v
==30795== ERROR SUMMARY: 0 errors from 0 contexts (suppressed: 0 from 0)