Make IghtBuffer Non{Mov,Copy}able
While making IghtBuffer non movable and non copyable, also
http::Request needs to be made non movable and non copyable.

IghtBuffer should not be movable and copyable because it
contains a pointer to a C structure.

It is not a big deal that http::Request is now non movable
and non copyable (it was on the TODO list already).