Add android traceroute code
This code was contributed by developers of Portolan, a network
measurement tool developed at University of Pisa.

Many many thanks!

I started merging this code to libight during the Tor winter dev
meeting in Valencia, after which I was able to compile it.

Then I worked on this together with @ValerioLuconi during a two day
hackfest held in Pisa (27-28 Apr).

Finally I also worked on this on May 21, at Nexa.

This code compiles and runs on Linux, an early version of it was
tested it both using IPv6 (at CNR), this version only on IPv4.

The original history of this contribution to libight counted 60+
commits that I have squashed to tidy things up.