Add android traceroute code
This code was contributed by developers of Portolan, a network
measurement tool developed at University of Pisa.

Many many thanks!

I started merging this code to libight during the Tor winter dev
meeting in Valencia, after which I was able to compile it.

This final diff is the result of further work by @ValerioLuconi and
me during a two day hackfest held in Pisa (27-28 Apr).

This code compiles and run on Linux, we have tested it both using
IPv4 and IPv6 (the latter was done at CNR premises).

The original history of this contribution to libight counted 60+
commits that I have squashed to tidy things up.