test/protocols/http: Avoid crash when compiled using g++
The problem arises when the code of a lambda function assigns
such lambda function to other code.

In such case, the closure bound to the lambda is destroyed
and this may lead to use after free access to data.

In particular, this has been tested when we compile libight
using g++ on Ubuntu 14.10.

I was already investigating this problem in the personal
repository that I use to study C++11; see:

- https://github.com/bassosimone/cxx-experiments/commit/b2a4d3a2

- https://github.com/bassosimone/cxx-experiments/commit/d0a42eee

- https://github.com/bassosimone/cxx-experiments/commit/03272f43

I should also create an issue explaining the problem addressed
by this commit and proposing a strategy to verify that the rest
of the code base is immune (or how to fix it if needed).