Reliability fix: event_add() event after each send()
The previous implementation triggered a timeout every 1.0 seconds
starting from the moment in which the object was created.

This was clearly wrong. First, because a timeout must be counted
starting from the last send. Second, because no timeout should
be issued if no send was issued.

Spotted during code review.