Detach thread before leaving thread function
With this fix committed, the code does not crash anymore when
we create another background thread.

Since I observed this both on MacOS and Linux with different
compilers and C++ libraries, it was probably a programmer error
that I not explicitly joined/detached the thread.

Debugging this issue I also noticed memory leaks on Linux
that I am going to investigate.