async.cpp: improve implementation
1) Delay destruction of tests that now are destroyed in a separate
loop after they have reached the final state and, importantly, after
the stack that led to the final state has been fully unwinded.

Should significantly reduce the risk of user-after-free and other
memory errors, plus it is morally equivalent to calling `delete` on
the test object after the full test has completed.

2) With the introduction of the pump() function now the code to
finalize the test is called in the threading context in which pump()
is called meaning that one could arrange to do so in the main
thread of execution, thereby reducing threading related concerns.

3) Introduce macros to simplify checking for consistency when
adding and/or removing stuff from tables.

4) While there, add few more comments and do some cosmetic changes.