Tweak .clang-format
I experimented with 2-spaces indent but I'm more comfortable
with 4-spaces. Also most code is already in 4-spaces.

I also like the possibility of putting short if statements
on a single line. IMHO it reduces the risk of doing this

     if (foo)
         something();
+        else();

by mistake because to add one more line you need to reformat
the code and go to the next line. At this point you will also
add the braces (at least this seems reasonable to me).