async: thread-safe handling of logs
This diff modifies async to implement thread-safe handling of
the logs produced by a test. Now start_test() receives one more
lambda function that is called when a test produced logs. When
a test is started, logs are routed to a map containing them,
using `test->on_log()`. When `pump()` is called, logs stored in
such map are passed to the above mentioned lambda.

The net results is that the App programmer does not need to
worry about thread safety issues, both for test-completion and
logs-available events, as long as she calls `pump()` from the
context of the main thread.