NOTABUG WONTFIX
Too much time was already spent trying to convince gcc-4.8 to generate
code coverage for us. The key to this is to use gcov-4.8 along with gcc-4.8,
yet coveralls fails because it cannot find some files. Guess this can be
fixed by teaching it about the directory where it should run (but then why
it does not fail for clang?).

Anyways, I'm going to be happy with measuring code coverage with clang
alone for now, and I am going to move on.

Here just strive for some more correctness by:

a) forcing gcov-4.6 with clang because this is the maximum version that is
   supposed to work, indeed gcov-4.8 leads to segfault when parsing coverage
   information generated by clang (also, I've read that for clang>=3.5 the
   coverage program to use is llvm-cov)

b) teaching coveralls to ignore src/ext to avoid measuring coverage of code
   that was not written by us

c) forcing version 4.8 of gcc, g++ and gcov in preparation of future changes
   where we'll be able to measure coverage also with gcc on travis-ci