.travis.yml: route gcov output to /dev/null
It seems there is no other way to silence off gcov, and note that we do
not care about that output because we have coveralls.

Hopefully, in case of error, gcov shall use the standard error which
is not muted and coveralls shall do the same.

Coveralls is a Python program and Python emits exceptions on the standard
error, hence filterting stdout is not going to prevent us from receiving
useful stacktraces in case of error.