Call the library libmeasurement-kit.la
The name of the project is MeasurementKit. The name of the repository
is measurement-kit (to avoid using camel case in the repository name,
following the convention I've seen used in the node.js world). The name
of the library is instead libmeasurement-kit.la (same as the name of the
repository but with lib in front and .la in the end). Every time it is not
possible to write measurement-kit, we write instead measurement_kit (this
is done, for example, in the include path because we want it to mimic the
namespaces and it is not possible to use the minus char in namespaces).

Triggered by the observation that it's counter-intuitive to see the
library using the underscore when the project uses the dash.