Rename and simplify namespaces
Use mk as namespace rather than measurement_kit. On this there
is agreement between me, @hellais (who originally prodded the idea),
and @AntonioLangiu.

Zap the common namespace. Prodded by @AntonioLangiu with me in
favour and @hellais agnostic with respect to this.

Removing the common namespace is good because stuff in such namespace
is nearly always pulled and having the namespace forces you either
to do `using namespace...` or to drop lots of `common::`.