poller: allow to modify default evdns-base state
I need this in the mobile applications. For example, with Android
there is no /etc/resolv.conf, and I need a mechanism to specify the
name servers to be used by the default evdns-base instance.

Note that in MeasurementKit there is also a custom Resolver object
constructed with default parameters, but this is not the same as the
evdns-base instance used by the Poller. The latter is, in fact, the
instance actually used when running, e.g., OONI tests.

More can be done to improve MeasurementKit with respect to the
poller and resolver facilities used when running tests, but this
is not the right moment to do so. We are close to a stable release
and hence we should make as minimal changes as possible.