Allow to clear DNS servers on Android
Previously I didn't notice that, when no /etc/resolv.conf is
found on the device, libevent sets 127.0.0.1 as resolver.

This is not what we want on Android. Because on Android there is
no resolver at 127.0.0.1. Hence add hooks to clear the configured
name servers, such that we can start from scratch.

While here, add a regress test case to check that what I would
like to do on Android (i.e. clear all nameservers and then force
a specific nameserver address) works.