test/http/stream.cpp: Fix error emitted to indicate EOF
Should fix the memory leak reported by @AntonioLangiu in #321.

This should not break other things in the tree, because the fix only
concerns the kind of error emitted in a test case.

After having investigated this issue, I have to say it is very
complex to reason about the code without the guarantee that what
you use is alive if you can use it. Especially with the usage
of smart pointers and lambda. Guess this is one more reason to
migrate to the static methods and shared pointers style with
which I've been experimenting for quite some time now.