Headers only implementation of OONI tests
For a couple of months now, the implementation of OONI tests is
composed of private headers and of private source files.

To increase readability of code, move bits of code from source
files to header files whenever possible.

Those header files are then included by the source files that are
compiled into the OONI sub-library.

Also, remove source files when they are empty after code was moved.