Start work to download dependencies on demand
When configure detects that a dependency is missing, it downloads
it and configure it. Next, when you make, the dependency is of course
built. And when you install it is installed.

Works thanks to wrapper configure.in and Makefile.in inside the
third_party/dependency directory. Configure.in downloads the dependency
using git and possibly applies patches, etc. Makefile.in forwards all
the targets to the real Makefile inside the dependency.